+++ src/alfasim_sdk/alfasim_sdk_api/api.h
@@ -621,6 +621,27 @@ DLL_EXPORT int get_deposition_thickness(
 );
 
 /*!
+    Logs a warning message at the solver logger.
+    Information will be logged as follow: "Warning: {warning_message}"
+
+    @param[in] ctx ALFAsim's plugins context.
+    @param[in] plugin_id ALFAsim's plugins id
+    @param[in] warning_message Warning Message.
+    @return An #error_code value.
+*/
+DLL_EXPORT int log_warning_message(void* ctx, const char* plugin_id, const char* warning_message);
+
+/*!
+    Logs an information message at the solver logger.
+
+    @param[in] ctx ALFAsim's plugins context.
+    @param[in] plugin_id ALFAsim's plugins id
+    @param[in] info_message Information Message.
+    @return An #error_code value.
+*/
+DLL_EXPORT int log_information_message(void* ctx, const char* plugin_id, const char* info_message);
+
+/*!
     Retrieves the tracer ID given a tracer reference. A tracer reference may be obtained from the
     user input data (See #get_plugin_input_data_reference API function for an example).